pyproject.toml
[project]
name = "catprinter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bleak==0.22.3",
    "numpy==2.2.4",
    "opencv-python==4.11.0.86",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "xkcd>=2.4.2",
]
